# Conflicts:
#	src/backends/dx/Common/Memory.h
#	src/backends/dx/Common/Pool.h
#	src/backends/dx/ShaderCompile/ASTEncoder.cpp
#	src/backends/dx/ShaderCompile/LuisaASTTranslator.cpp
#	src/backends/dx/WorldManagement/World.h

+++ src/ast/function_builder.cpp
@@ -260,8 +260,8 @@ void FunctionBuilder::for_(const Statement *init, const Expression *condition, c
 }
 
 void FunctionBuilder::mark_variable_usage(uint32_t uid, Usage usage) noexcept {
-    auto old_usage = _variable_usages[uid];
-    _variable_usages[uid] = static_cast<Usage>(old_usage | usage);
+    auto old_usage = static_cast<uint32_t>(_variable_usages[uid]);
+    _variable_usages[uid] = static_cast<Usage>(old_usage | static_cast<uint32_t>(usage));
 }
 
 FunctionBuilder *FunctionBuilder::create(Function::Tag tag) noexcept {
